Extract query, msg to dm
