+++ src/slash-commands/dm.ts
@@ -23,6 +23,11 @@ import { guessAndSetDMRoom } from "../Rooms";
 import { TimelineRenderingType } from "../contexts/RoomContext";
 import { Command } from "./command";
 import { CommandCategories } from "./interface";
+import LegacyCallHandler from "../LegacyCallHandler";
+import { ensureDMExists } from "../createRoom";
+import dis from "../dispatcher/dispatcher";
+import { ViewRoomPayload } from "../dispatcher/payloads/ViewRoomPayload";
+import { Action } from "../dispatcher/actions";
 
 export const converttodm = new Command({
     command: "converttodm",
@@ -48,4 +53,77 @@ export const converttoroom = new Command({
         return success(guessAndSetDMRoom(room, false));
     },
     renderingTypes: [TimelineRenderingType.Room],
+});
+
+export const query = new Command({
+    command: "query",
+    description: _td("Opens chat with the given user"),
+    args: "<user-id>",
+    runFn: function (cli, roomId, threadId, userId) {
+        // easter-egg for now: look up phone numbers through the thirdparty API
+        // (very dumb phone number detection...)
+        const isPhoneNumber = userId && /^\+?[0123456789]+$/.test(userId);
+        if (!userId || ((!userId.startsWith("@") || !userId.includes(":")) && !isPhoneNumber)) {
+            return reject(this.getUsage());
+        }
+
+        return success(
+            (async (): Promise<void> => {
+                if (isPhoneNumber) {
+                    const results = await LegacyCallHandler.instance.pstnLookup(userId);
+                    if (!results || results.length === 0 || !results[0].userid) {
+                        throw new UserFriendlyError("Unable to find Matrix ID for phone number");
+                    }
+                    userId = results[0].userid;
+                }
+
+                const roomId = await ensureDMExists(cli, userId);
+                if (!roomId) throw new Error("Failed to ensure DM exists");
+
+                dis.dispatch<ViewRoomPayload>({
+                    action: Action.ViewRoom,
+                    room_id: roomId,
+                    metricsTrigger: "SlashCommand",
+                    metricsViaKeyboard: true,
+                });
+            })(),
+        );
+    },
+    category: CommandCategories.actions,
+});
+
+export const msg = new Command({
+    command: "msg",
+    description: _td("Sends a message to the given user"),
+    args: "<user-id> [<message>]",
+    runFn: function (cli, roomId, threadId, args) {
+        if (args) {
+            // matches the first whitespace delimited group and then the rest of the string
+            const matches = args.match(/^(\S+?)(?: +(.*))?$/s);
+            if (matches) {
+                const [userId, msg] = matches.slice(1);
+                if (userId && userId.startsWith("@") && userId.includes(":")) {
+                    return success(
+                        (async (): Promise<void> => {
+                            const roomId = await ensureDMExists(cli, userId);
+                            if (!roomId) throw new Error("Failed to ensure DM exists");
+
+                            dis.dispatch<ViewRoomPayload>({
+                                action: Action.ViewRoom,
+                                room_id: roomId,
+                                metricsTrigger: "SlashCommand",
+                                metricsViaKeyboard: true,
+                            });
+                            if (msg) {
+                                cli.sendTextMessage(roomId, msg);
+                            }
+                        })(),
+                    );
+                }
+            }
+        }
+
+        return reject(this.getUsage());
+    },
+    category: CommandCategories.actions,
 });
